Refactor intelligent response capture into modular classes
